Fix missing openconfig-models dependency in simulator runtime
The openconfig-models dependency was incorrectly scoped as 'test' in the
odl-gnmi-device-simulator pom.xml. This caused the maven-dependency-plugin
to exclude the JAR from the generated 'lib' directory used by the
standalone application.

As a result, the ServiceLoader could not locate the YangModelBindingProvider
implementations at runtime, causing the simulator to fail during startup
with an EffectiveModelContextBuilderException.

Change-Id: Ida1df44b4c2708a8d4504f7b6e6058db9f2de917
Signed-off-by: tobias.pobocik <tobias.pobocik@pantheon.tech>