Fix gNMI integration tests
Resolve multiple issues with wrong paths, wrong initialization
and creation of GnmiSouthboundModule to register RPCs to of
test modules. The port also is hard coded to 8181 instead of
finding open one.
Also remove ability to connect device without paraphrase field
since even when this field is not set, it should be included.

Change-Id: I1fb41b6301a5c1b32846963fcb05a95d2cb1d06c
Signed-off-by: tobias.pobocik <tobias.pobocik@pantheon.tech>