Clarify occasional test failures
connectMultipleDevicesTest is our occasionally failing test. It differs
from others in that it configures/writes to datastore multiple devices.

lighty.io has been using full RNC app to handle this testcase, in ODL
we have in-memory datastore based tests.

There are multiple reasons why we fail now:
- we have to clear datastore in between tests
- we have to improve threading of our test

For this to clarify and isolate separate multiple devices connect and
disconnect test cases to its own class.

In case of failure we get similar error as in NETCONF-1414 which can
be causing also not READY error.

Change-Id: I6541a5ba7f979ec25f165f91d19b59ad502b2272
Signed-off-by: Ivan Hrasko <ivan.hrasko@pantheon.tech>